chore: move and rename some files
